bot/pyproject.toml
[tool.poetry]
name = "tags-bot"
version = "1.0.0"
description = "A bot that helps with previewing your Meta/OpenGraph tags before you use them on your site."
authors = ["spifory <luvuyosilubane50@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
hikari = {extras = ["server"], version = "^2.0.0.dev120"}
hikari-crescent = "^0.6.2"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.284"
black = "^23.7.0"

[tool.ruff]
select = ["ALL"]
ignore = ["ANN101", "ANN102"]
max-line-length = 100

[tool.ruff.per-file-ignores]
"__init__.py" = ["F403"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
